move buffer pool
